+++ packages/connect/e2e/__fixtures__/ethereumGetPublicKey.ts
@@ -13,6 +13,7 @@ const generatedTests = commonFixtures.tests.flatMap(({ parameters, result }) => 
         chainCode: result.chain_code,
         publicKey: result.public_key,
         xpub: result.xpub,
+        displayablePublicKey: result.xpub,
     },
 }));